requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
plotly>=5.18.0
requests>=2.31.0
pydeck>=0.8.0
